+++ tests/Unit/Console/LcbApplicationTest.php
@@ -14,6 +14,6 @@ final class LcbApplicationTest extends TestCase
         $application = new LcbApplication();
 
         self::assertSame('Laravel Cloud Blueprint', $application->getName());
-        self::assertSame('0.1.0-alpha.3', $application->getVersion());
+        self::assertSame('0.1.0-alpha.4', $application->getVersion());
     }
 }